+++ test/libsolidity/syntaxTests/inlineAssembly/slotnum.sol
@@ -1,0 +1,10 @@
+contract C {
+    function f() public view returns (uint64 ret) {
+        assembly {
+            ret := slotnum()
+        }
+    }
+}
+// ====
+// EVMVersion: >=amsterdam
+// ----