+++ test/libsolidity/syntaxTests/types/magic_block_slotnum_amsterdam.sol
@@ -1,0 +1,13 @@
+contract C {
+    function f() public view returns (uint64) {
+        return block.slotnum;
+    }
+    function g() public view returns (uint64 ret) {
+        assembly {
+            ret := slotnum()
+        }
+    }
+}
+// ====
+// EVMVersion: >=amsterdam
+// ----