+++ test/libsolidity/syntaxTests/viewPureChecker/slotnum_view.sol
@@ -1,0 +1,13 @@
+contract C {
+    function f() external view returns (uint64 a) {
+        assembly {
+            a := slotnum()
+        }
+    }
+    function g() external view returns (uint64) {
+        return block.slotnum;
+    }
+}
+// ====
+// EVMVersion: >=amsterdam
+// ----